requirements.txt
pandas
sqlalchemy
mysql-connector-python
matplotlib
seaborn
jupyter
python-dotenvpip